Fastapi/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart

numpy
scipy
python-dotenv
pandas
h5py
tensorflow
Keras
Keras-Preprocessing
tqdm
Pillow